requirements.txt
pydantic
requests
uuid
gTTS
zhon
jieba
azure-cognitiveservices-speech
pymongo
python-dotenv
azure-identity
azure-storage-blob
matplotlib
streamlit
streamlit-elements
streamlit-mic-recorder
werkzeug
faker
bs4
google-cloud-aiplatform
google-cloud-storage
google-generativeai
google-cloud-translate
google-cloud-texttospeech
cryptography
